ran linter

+++ internal/proto6server/server_readstatebytes.go
@@ -88,7 +88,7 @@ func (s *Server) ReadStateBytes(ctx context.Context, proto6Req *tfprotov6.ReadSt
 				chunk := tfprotov6.ReadStateByteChunk{
 					StateByteChunk: tfprotov6.StateByteChunk{
 						Bytes:       readBytes[:byteCount],
-						TotalLength: int64(totalLength),
+						TotalLength: totalLength,
 						Range: tfprotov6.StateByteRange{
 							Start: int64(rangeStart),
 							End:   int64(rangeStart + byteCount),
